exponential-smoothing-predictor/src/requirements.txt
python-slugify==8.0.1
jproperties==2.1.1
requests==2.31.0
msgpack==1.0.7
numpy==1.26.3
pandas==2.1.4
python-dotenv==1.0.0
python-qpid-proton==0.39.0
influxdb-client==1.39.0